Add EditPost to manifest

+++ ReThrift/app/src/main/java/rethrift/rethrift/MyPostsActivity.java
@@ -128,7 +128,7 @@ public class MyPostsActivity extends AppCompatActivity {
                         posts.add(
                                 new Post(postJson.getInt("id"),
                                          postJson.getString("title"),
-                                         "$" + postJson.getDouble("price"),
+                                         "$" + postJson.getInt("price"),
                                          postJson.getString("state"),
                                          postJson.getDouble("latitude"),
                                          postJson.getDouble("longitude"),
