+++ ReThrift/app/src/main/java/rethrift/rethrift/WatchListActivity.java
@@ -127,7 +127,7 @@ public class WatchListActivity extends AppCompatActivity {
                         posts.add(
                                 new Post(postJson.getInt("id"),
                                          postJson.getString("title"),
-                                         "$" + postJson.getDouble("price"),
+                                         "$" + postJson.getInt("price"),
                                          postJson.getString("state"),
                                          postJson.getDouble("latitude"),
                                          postJson.getDouble("longitude"),
